requirements.txt
pillow==10.4.0
PyMuPDF==1.24.7
pytesseract==0.3.10
PyPDF2==3.0.1